Fixes to evaluate functions and dl
Decision level of a variable or value is now set to when it gets a
value assigned, not when the literal is assigned. This prevents issues
where a literal is assigned at some decision level but due to its
depencies on other variables it can't get a value until later.
Then, the backtracking would reset the variable correctly.

Evaluate functions now assume that a value can only be set once.